requirements.txt
streamlit
openai>=1.0.0
python-docx
fpdf
python-dotenv